pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jml-engine"
dynamic = ["version"]
description = "IAM Lifecycle Automation Engine for Joiner-Mover-Leaver workflows across AWS, Azure, and SaaS applications"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "JML Engine Team", email = "team@jml-engine.dev"}
]
maintainers = [
    {name = "JML Engine Team", email = "team@jml-engine.dev"}
]
keywords = ["iam", "identity", "automation", "joiner-mover-leaver", "aws", "azure", "github", "google-workspace", "slack", "compliance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.104.1,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "streamlit>=1.28.1,<2.0.0",
    "boto3>=1.34.0,<2.0.0",
    "azure-identity>=1.15.0,<2.0.0",
    "azure-mgmt-authorization>=4.0.0,<5.0.0",
    "azure-mgmt-resource>=23.0.0,<24.0.0",
    "PyGitHub>=2.1.1,<3.0.0",
    "google-api-python-client>=2.108.0,<3.0.0",
    "google-auth>=2.25.0,<3.0.0",
    "google-auth-oauthlib>=1.2.0,<2.0.0",
    "slack-sdk>=3.27.0,<4.0.0",
    "pandas>=2.1.0,<3.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "structlog>=23.2.0,<24.0.0",
    "python-json-logger>=2.0.7,<3.0.0",
    "click>=8.1.7,<9.0.0",
    "rich>=13.7.0,<14.0.0",
    "cryptography>=41.0.7,<42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-xdist>=3.0.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.0.0,<2.0.0",
    "pre-commit>=3.0.0,<4.0.0",
    "httpx>=0.25.0,<1.0.0",
    "bandit>=1.7.0,<2.0.0",
    "safety>=2.0.0,<3.0.0",
    "commitizen>=3.0.0,<4.0.0",
    "build>=1.0.0,<2.0.0",
    "twine>=4.0.0,<5.0.0",
]
docs = [
    "mkdocs>=1.5.0,<2.0.0",
    "mkdocs-material>=9.0.0,<10.0.0",
    "mkdocstrings>=0.22.0,<1.0.0",
    "mkdocstrings-python>=1.0.0,<2.0.0",
]

[project.scripts]
jmlctl = "jml_engine.cli.jmlctl:main"

[project.urls]
Homepage = "https://github.com/yourusername/jml-engine"
Documentation = "https://github.com/yourusername/jml-engine#readme"
Repository = "https://github.com/yourusername/jml-engine"
Issues = "https://github.com/yourusername/jml-engine/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["jml_engine*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 100
target-version = "py38"
[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]  # assert false

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--verbose",
    "--tb=short",
    "--cov=jml_engine",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["jml_engine"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
